Move `extract_discovered` to `MdnsResponse`
This results in better encapsulation.
